New upgrade code (fixes #11)

+++ install.sh
@@ -64,12 +64,48 @@ if [ $makerfaire2018 -eq 1 ]; then
   fi
 fi
 
-if [ $travis_chroot -eq 0 -a ! -e "/dev/ear0" ]; then
-  echo "Please install ears driver https://github.com/pguyot/tagtagtag-ears"
-  exit 1
+if [ $upgrade -eq 1 -a $makerfaire2018 -eq 0 -a -d /home/pi/wm8960 ]; then
+  echo "Updating sound driver" > /tmp/pynab.upgrade
+  cd /home/pi/wm8960
+  pull=`git pull`
+  if [ "$pull" != "Already up to date." ]; then
+    make && sudo make install
+    sudo touch /tmp/pynab.upgrade.reboot
+  fi
+else
+  if [ $travis_chroot -eq 0 -a ! -e "/dev/ear0" ]; then
+    echo "Please install ears driver https://github.com/pguyot/tagtagtag-ears"
+    exit 1
+  fi
+fi
+
+if [ $upgrade -eq 1 ]; then
+  echo "Updating ears driver" > /tmp/pynab.upgrade
+  if [ -d /home/pi/tagtagtag-ears ]; then
+    cd /home/pi/tagtagtag-ears
+    pull=`git pull`
+    if [ "$pull" != "Already up to date." ]; then
+      make && sudo make install
+      sudo touch /tmp/pynab.upgrade.reboot
+    fi
+  else
+    cd /home/pi
+    git clone https://github.com/pguyot/tagtagtag-ears
+    make && sudo make install
+    sudo touch /tmp/pynab.upgrade.reboot
+  fi
+else
+  if [ $travis_chroot -eq 0 -a ! -e "/dev/ear0" ]; then
+    echo "Please install ears driver https://github.com/pguyot/tagtagtag-ears"
+    exit 1
+  fi
 fi
 
 if [ $makerfaire2018 -eq 0 ]; then
+  if [ $upgrade -eq 1 ]; then
+    echo "Updating ASR models" > /tmp/pynab.upgrade
+  fi
+
   # maker faire card has no mic, no need to install kaldi
   if [ ! -d "/opt/kaldi" ]; then
     echo "Installing precompiled kaldi into /opt"
@@ -90,6 +126,7 @@ if [ $makerfaire2018 -eq 0 ]; then
   fi
 fi
 
+cd $root_dir
 if [ ! -d "venv" ]; then
   if ! [ -x "$(command -v python3.7)" ] ; then
     echo "Please install Python 3.7 (you might need to upgrade your Raspbian distribution)"
@@ -101,11 +138,18 @@ if [ ! -d "venv" ]; then
 fi
 
 echo "Installing PyPi requirements"
+if [ $upgrade -eq 1 ]; then
+  echo "Updating Python requirements" > /tmp/pynab.upgrade
+fi
 # Start with wheel which is required to compile some of the other requirements
 venv/bin/pip install wheel
 venv/bin/pip install -r requirements.txt
 
 if [ $makerfaire2018 -eq 0 ]; then
+  if [ $upgrade -eq 1 ]; then
+    echo "Updating NLU models" > /tmp/pynab.upgrade
+  fi
+
   # maker faire card has no mic, no need to install snips
   if [ ! -d "venv/lib/python3.7/site-packages/snips_nlu_fr" ]; then
     echo "Downloading snips_nlu models for French"
@@ -151,13 +195,21 @@ if [ $trust -ne 1 ]; then
   fi
 fi
 
-if [ ! -e '/etc/nginx/sites-enabled/pynab' ]; then
-  echo "Installing nginx configuration file"
-  if [ -h '/etc/nginx/sites-enabled/default' ]; then
-    sudo rm /etc/nginx/sites-enabled/default
+if [ $upgrade -eq 0 ]; then
+  if [ ! -e '/etc/nginx/sites-enabled/pynab' ]; then
+    echo "Installing nginx configuration file"
+    if [ -h '/etc/nginx/sites-enabled/default' ]; then
+      sudo rm /etc/nginx/sites-enabled/default
+    fi
+    sudo cp nabweb/nginx-site.conf /etc/nginx/sites-enabled/pynab
+    if [ $travis_chroot -eq 0 ]; then
+      sudo systemctl restart nginx
+    fi
   fi
-  sudo cp nabweb/nginx-site.conf /etc/nginx/sites-enabled/pynab
-  if [ $travis_chroot -eq 0 ]; then
+else
+  echo "Restarting Nginx" > /tmp/pynab.upgrade
+  if [ -e '/etc/nginx/sites-enabled/pynab' ]; then
+    sudo cp nabweb/nginx-site.conf /etc/nginx/sites-enabled/pynab
     sudo systemctl restart nginx
   fi
 fi
@@ -169,10 +221,15 @@ psql -U pynab -c '' 2>/dev/null || {
   sudo -u postgres psql -U postgres -c "ALTER ROLE pynab CREATEDB"
 }
 
+if [ $upgrade -eq 1 ]; then
+  echo "Updating data models" > /tmp/pynab.upgrade
+fi
 venv/bin/python manage.py migrate
+
 if [ $upgrade -eq 0 ]; then
   venv/bin/django-admin compilemessages
 else
+  echo "Updating localization messages" > /tmp/pynab.upgrade
   for module in nab*/locale; do
     (
       cd `dirname ${module}`
@@ -191,6 +248,9 @@ if [ $travis_chroot -eq 1 ]; then
 fi
 
 # copy service files
+if [ $upgrade -eq 1 ]; then
+  echo "Installing service files" > /tmp/pynab.upgrade
+fi
 for service_file in nabd/nabd.socket */*.service ; do
   name=`basename ${service_file}`
   sudo sed -e "s|/home/pi/pynab|${root_dir}|g" < ${service_file} > /tmp/${name}
@@ -199,17 +259,32 @@ for service_file in nabd/nabd.socket */*.service ; do
   sudo systemctl enable ${name}
 done
 
-if [ $travis_chroot -eq 0 ]; then
-  sudo systemctl start nabd.socket
-  sudo systemctl start nabd.service
-
-  # start services
-  for service_file in */*.service ; do
-    name=`basename ${service_file}`
-    if [ "${name}" != "nabd.service" -a "${name}" != "nabweb.service" ]; then
-      sudo systemctl start ${name}
+if [ -e /tmp/pynab.upgrade.reboot ]; then
+  echo "Upgrade requires reboot, rebooting now"
+  sudo rm -f /tmp/pynab.upgrade
+  sudo rm -f /tmp/pynab.upgrade.reboot
+  sudo reboot
+else
+  if [ $travis_chroot -eq 0 ]; then
+    if [ $upgrade -eq 1 ]; then
+      echo "Restarting services" > /tmp/pynab.upgrade
     fi
-  done
+    sudo systemctl start nabd.socket
+    sudo systemctl start nabd.service
 
-  sudo systemctl start nabweb.service
+    # start services
+    for service_file in */*.service ; do
+      name=`basename ${service_file}`
+      if [ "${name}" != "nabd.service" -a "${name}" != "nabweb.service" ]; then
+        sudo systemctl start ${name}
+      fi
+    done
+
+    if [ $upgrade -eq 1 ]; then
+      echo "Restarting website" > /tmp/pynab.upgrade
+      sudo systemctl restart nabweb.service
+    else
+      sudo systemctl start nabweb.service
+    fi
+  fi
 fi
